factor out a template for testing file based data container tables
